requirements.txt
certifi==2018.10.15
chardet==3.0.4
idna==2.7
paho-mqtt==1.4.0
python-dotenv==0.9.1
requests==2.20.0
urllib3==1.24
